Ajnabee-server/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Django==3.1.7
djangorestframework==3.12.4
djutil==0.2
filelock==3.0.12
gunicorn==20.1.0
idna==2.10
importlib-metadata==3.10.0
joblib==1.0.1
nltk==3.5
numpy==1.20.2
packaging==20.9
pandas==1.2.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
regex==2021.3.17
requests==2.25.1
sacremoses==0.0.44
scikit-learn==0.24.1
scipy==1.6.2
sentence-transformers==1.0.4
sentencepiece==0.1.95
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.1.0
tokenizers==0.10.1
torch==1.8.1
tqdm==4.59.0
transformers==4.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
zipp==3.4.1
